Initial pass at Web Audio vs HTML5 detection
Still a little dirty, but gets the job done. Also is a start at
namespacing.

This appears to be working across browsers, but now I’m not getting any
audio on iOS (Chrome/Safari/Opera), though the behaviors look good
(page loads, button text displays and updates, and I get a loading
spinner when buffering). I swear it was working before, but will double
check that.