Working Web Audio across mobile and desktop
Verified in Sarari and Chrome desktop, and Safari, Chrome, and Opera
mobile. Just a dirty replacement of my code with a sample from
matt-harrison.com plus some tweaks to address errors and get it
working. Need to run through and figure out why this is working and my
original code is not.